+++ R/models.R
@@ -54,8 +54,8 @@ list_available_models.huggingface <- function(service) {
 #' @export
 list_available_models.anthropic <- function(service) {
   c(
-    "claude-3-opus-20240229", "claude-3-sonnet-20240229", "claude-3-haiku-20240307",
-    "claude-2.1", "claude-instant-1.2"
+    "claude-3-5-sonnet-20240620", "claude-3-opus-20240229", "claude-3-sonnet-20240229",
+    "claude-3-haiku-20240307", "claude-2.1", "claude-instant-1.2"
   )
 }