+++ NEWS.md
@@ -1,6 +1,7 @@
 # gptstudio (development version)
 
 - Fixed a bug that showed the message "ChatGPT responded" even when other service was being used in "Chat in source" related addins. #213
+- Added claude-3.5-sonnet model from Anthropic.
 
 ## gptstudio 0.4.0